Added third party support
